# apps/subtitles/internal/server/api_openapi.json
{
  "openapi": "3.1.0",
  "info": {"title": "Kinosail Subtitles API", "version": "1.0.0", "description": "Versioned local subtitle automation API."},
  "servers": [{"url": "/"}],
  "security": [{"bearerAuth": []}],
  "components": {
    "schemas": {"SubtitleEdit":{"type":"object","additionalProperties":false,"required":["language"],"properties":{"role":{"type":"string","enum":["translation","captions"],"description":"User-confirmed role. Generated dialogue defaults to translation, not complete accessibility captions."},"draftId":{"type":"string","pattern":"^[a-f0-9]{64}$","description":"Ready local draft bound to this video and language; mutually exclusive with data."},"text":{"type":"string","maxLength":4194304,"description":"Optional corrected UTF-8 SRT text, at most 4 MiB. Retains the source original."},"language":{"type":"string","minLength":2,"maxLength":64},"data":{"type":"string","contentEncoding":"base64","maxLength":5592408,"description":"Original subtitle bytes, at most 4 MiB before base64 encoding. Omit to edit the installed subtitle."},"fingerprint":{"type":"string","pattern":"^(missing|[a-f0-9]{64})$","description":"Current fingerprint returned by inspect or preview. Required when saving."},"encoding":{"type":"string","enum":["auto","utf-8","utf-16le","utf-16be","windows-1252","windows-1250","windows-1251","windows-1253","windows-1254","windows-1255","windows-1256","windows-874","shift-jis","gb18030","big5","euc-kr"]},"offsetMilliseconds":{"type":"integer","minimum":-120000,"maximum":120000},"automaticSync":{"type":"boolean"},"removeCredits":{"type":"boolean","default":false},"mergeRepeated":{"type":"boolean","default":false},"anchors":{"type":"array","maxItems":8,"items":{"type":"object","additionalProperties":false,"required":["atMilliseconds","offsetMilliseconds"],"properties":{"atMilliseconds":{"type":"integer","minimum":0,"maximum":129600000},"offsetMilliseconds":{"type":"integer","minimum":-120000,"maximum":120000}}}}}}},
    "securitySchemes": {"bearerAuth": {"type": "http", "scheme": "bearer"}},
    "parameters": {
      "Id": {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}},
...
    },
    "responses": {"Result": {"description": "Successful response; the operation schema is versioned with the endpoint."}}
  },
  "paths": {
...
    "/api/v1/subtitle-library/{id}/draft": {"parameters":[{"$ref":"#/components/parameters/Id"}],"get":{"operationId":"getLocalSubtitleDraft","description":"Read the current local draft, recognition confidence, and estimated word timing.","parameters":[{"name":"language","in":"query","schema":{"type":"string"}}],"responses":{"default":{"$ref":"#/components/responses/Result"}}},"post":{"operationId":"generateLocalSubtitleDraft","description":"Start or cancel local bitmap OCR or Whisper transcription. One bounded job and one volatile draft are retained; nothing is installed automatically.","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"required":["language","method","action"],"properties":{"language":{"type":"string"},"method":{"enum":["ocr","transcription"]},"action":{"enum":["start","cancel"]},"draftId":{"type":"string","pattern":"^[a-f0-9]{64}$","description":"Required only for cancellation."}}}}}},"responses":{"default":{"$ref":"#/components/responses/Result"}}}},
    "/api/v1/subtitle-library/{id}/inspect": {"parameters":[{"$ref":"#/components/parameters/Id"}],"get":{"operationId":"inspectSubtitle","parameters":[{"name":"language","in":"query","schema":{"type":"string","minLength":2,"maxLength":64}}],"responses":{"default":{"$ref":"#/components/responses/Result"}}}},
    "/api/v1/subtitle-library/{id}/export": {"parameters":[{"$ref":"#/components/parameters/Id"}],"get":{"operationId":"exportSubtitle","parameters":[{"name":"language","in":"query","schema":{"type":"string","minLength":2,"maxLength":64}},{"name":"format","in":"query","schema":{"type":"string","enum":["srt","vtt","original"],"default":"srt"}}],"responses":{"200":{"description":"Subtitle download; same-format WebVTT and original preserve source bytes."},"default":{"$ref":"#/components/responses/Result"}}}},
    "/api/v1/subtitle-library/{id}/preview": {"parameters":[{"$ref":"#/components/parameters/Id"}],"post":{"operationId":"previewSubtitleEdit","description":"Preview conversion and timing without changing the sidecar.","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubtitleEdit"}}}},"responses":{"default":{"$ref":"#/components/responses/Result"}}}},
    "/api/v1/subtitle-library/{id}/apply": {"parameters":[{"$ref":"#/components/parameters/Id"}],"post":{"operationId":"applySubtitleEdit","description":"Save a reviewed subtitle using a required current fingerprint, original archive, recovery copy, and automatic-upgrade protection.","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubtitleEdit"}}}},"responses":{"default":{"$ref":"#/components/responses/Result"}}}},
    "/api/v1/subtitle-library/{id}/audio": {"parameters":[{"$ref":"#/components/parameters/Id"}],"post":{"operationId":"analyzeSubtitleAudio","description":"Analyze main dialogue locally; return bounded waveform and speech probability bins.","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"properties":{"language":{"type":"string","minLength":2,"maxLength":64}}}}}},"responses":{"default":{"$ref":"#/components/responses/Result"}}}},
...
    "/api/v1/metadata/bulk": {"post": {"operationId": "bulkEditMetadata", "security": [{"bearerAuth": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["ids"], "properties": {"ids": {"type": "array", "minItems": 1, "maxItems": 100, "uniqueItems": true, "items": {"type": "string", "pattern": "^[a-f0-9]{16}$"}}, "title": {"type": "string", "maxLength": 200}, "year": {"type": "string", "maxLength": 4}, "plot": {"type": "string", "maxLength": 5000}, "rating": {"type": "string", "maxLength": 32}, "tagline": {"type": "string", "maxLength": 300}, "genres": {"type": "string", "maxLength": 500}}}}}}, "responses": {"default": {"$ref": "#/components/responses/Result"}}}},
...
  }
}